+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,7 @@
+{
+  "ExpandedNodes": [
+    ""
+  ],
+  "SelectedNode": "\\C:\\Users\\franc\\Raccolta\\SRS",
+  "PreviewInSolutionExplorer": false
+}
