--- README.md
@@ -1,7 +1,0 @@
-## Truthster has been prepared with some integrated packages:
-
-1) Truffle
-2) Bootstrap + Matarial components
-3) Ethers
-
-Clone this repository in order to have a ready to start project for Web3 dApps.
