+++ poison-essence/.vs/VSWorkspaceState.json
@@ -1,0 +1,7 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\src"
+  ],
+  "PreviewInSolutionExplorer": false
+}
